tls: split out common, openssl and mbedtls code
 - introduce lib/tls/mbedtls lib/tls/openssl
 - move wrapper into lib/tls/mbedtls/wrapper
 - introduce private helpers to hide backend

This patch doesn't replace or remove the wrapper, it moves it
to lib/tls/mbedtls/wrapper.

But it should be now that the ONLY functions directly consuming
wrapper apis are isolated in

  - lib/tls/mbedtls/client.c (180 lines)
  - lib/tls/mbedtls/server.c (317 lines)
  - lib/tls/mbedtls/ssl.c    (325 lines)

In particular there are no uses of openssl or mbedtls-related
constants outside of ./lib/tls any more.

+++ plugins/protocol_lws_sshd_demo.c
@@ -27,6 +27,7 @@
 #include <lws-ssh.h>
 
 #include <string.h>
+#include <stdlib.h>
 
 #define TEST_SERVER_KEY_PATH "/etc/lws-test-sshd-server-key"
 
@@ -160,7 +161,7 @@ ssh_ops_tx(void *_priv, int stdch, uint8_t *buf, size_t len)
 	if (stdch != LWS_STDOUT)
 		return 0;
 
-	if (priv->len - priv->pos < chunk)
+	if ((size_t)(priv->len - priv->pos) < chunk)
 		chunk = priv->len - priv->pos;
 
 	if (!chunk)
